refactor(cli): move describeAuthError into a shared authmsg package
internal/cli/app cannot import internal/cli/login (login -> internal/cli/cmd
-> app is an import cycle), so relocate the mapper to internal/cli/authmsg,
whose only dependency is pkg/auth, and have login call it instead of keeping
its own copy. User-facing strings are unchanged.

Also add the a.PrintBanner() call that every other top-level command makes
before doing its work, which login and logout were missing.
